+++ brutils/__init__.py
@@ -22,7 +22,4 @@ from brutils.phone import (
     is_valid as is_valid_phone,
 )
 from brutils.email import is_valid_email as is_valid_email
-from brutils.license_plate import (
-    is_valid_old_format as is_valid_license_plate_old_format,
-    is_valid_mercosul as is_valid_license_plate_mercosul,
-)
+from brutils.license_plate import is_valid_mercosul as is_valid_license_plate_mercosul