+++ tests/test_license_plate.py
@@ -1,7 +1,7 @@
 # from unittest.mock import patch
 
 # from brutils.phone import is_valid_landline, is_valid_mobile, is_valid
-from brutils.license_plate import is_valid_mercosul, is_valid_old_format
+from brutils.license_plate import is_valid_mercosul
 
 from unittest import TestCase, main
 
@@ -36,36 +36,6 @@ class TestLicensePlate(TestCase):
 
         # Check if function is case insensitive
         self.assertIs(is_valid_mercosul("abc4e67"), True)
-
-    def test_is_valid_license_plate_old_format(self):
-        # When license plate is not string, returns False
-        self.assertIs(is_valid_old_format(1234567), False)
-
-        # When license plate doesn't match the pattern LLLNNNN, returns False
-        self.assertIs(is_valid_old_format("ABCDEFG"), False)
-        self.assertIs(is_valid_old_format("1234567"), False)
-        self.assertIs(is_valid_old_format("ABC123"), False)
-        self.assertIs(is_valid_old_format("ABCD234"), False)
-        self.assertIs(is_valid_old_format("ABC12F4"), False)
-        self.assertIs(is_valid_old_format("ABC123G"), False)
-        self.assertIs(is_valid_old_format("ABC123"), False)
-
-        # When license plate is an empty string, returns False
-        self.assertIs(is_valid_old_format(""), False)
-
-        # When license plate's length is different of 7, returns False
-        self.assertIs(is_valid_old_format("ABC12345"), False)
-
-        # When license plate has separator, returns false
-        self.assertIs(is_valid_old_format("ABC-1234"), False)
-
-        # When license plate is valid
-        self.assertIs(is_valid_old_format("ABC4567"), True)
-        self.assertIs(is_valid_old_format("AAA1111"), True)
-        self.assertIs(is_valid_old_format("XXX9999"), True)
-
-        # Check if function is case insensitive
-        self.assertIs(is_valid_old_format("abc4567"), True)
 
 
 if __name__ == "__main__":